+++ scripts/hosts.js
@@ -1,3 +1,4 @@
+export const mycroftUrl = 'https://mycroftproject.com/installos.php/';
 export const googleReverseImageSearchUrl = 'https://www.google.com/searchbyimage?sbisrc=1&safe=off&image_url=';
 export const googleLensUrl = 'https://lens.google.com/uploadbyurl?url=';
 export const chatGPTUrl = 'https://chatgpt.com/';
@@ -5,3 +6,6 @@ export const googleAIStudioUrl = 'https://aistudio.google.com/app/prompts/new_ch
 export const perplexityAIUrl = 'https://www.perplexity.ai/';
 export const poeUrl = 'https://poe.com';
 export const claudeUrl = 'https://claude.ai';
+export const youUrl = 'https://you.com';
+export const andiUrl = 'https://andisearch.com';
+export const exaUrl = 'https://exa.ai/search';